src/main/java/com/foxminded/university/service/TeacherService.java
package com.foxminded.university.service;

import com.foxminded.university.dao.TeacherDao;
import com.foxminded.university.domain.Teacher;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import java.util.List;

@Component
public class TeacherService implements Service<Teacher>{

    private final TeacherDao teacherDao;

    @Autowired
    public TeacherService(TeacherDao teacherDao) {
        this.teacherDao = teacherDao;
    }

    @Override
    public Teacher getById(int id) {
        return teacherDao.getById(id);
    }

    @Override
    public List<Teacher> getAll() {
        return teacherDao.getAll();
    }

    @Override
    public void save(Teacher teacher) {
        teacherDao.save(teacher);
    }

    @Override
    public void update(Teacher teacher) {
        teacherDao.update(teacher);
    }

    @Override
    public void delete(int id) {
        teacherDao.delete(id);
    }
}
